backend/pyproject.toml
[project]
name = "smart-radar-backend"
version = "1.0.0"
description = "SMART RADAR MVP Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn==0.24.0",
    "motor==3.3.2",
    "pymongo==4.6.0",
    "websockets==12.0",
    "pydantic==2.5.0",
    "python-multipart==0.0.6",
    "python-dotenv==1.0.0",
    "openai==1.3.7",
    "httpx==0.25.2",
    "celery==5.3.4",
    "redis==5.0.1",
    "aiohttp>=3.12.15",
    "google>=3.0.0",
    "google-generativeai>=0.8.5",
    "scikit-learn>=1.7.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
